+++ packages/orq-rc/RELEASES.md
@@ -9619,3 +9619,13 @@ Based on:
 - [typescript v4.10.0-rc.64] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.64] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.64 - packages/orq-rc
+
+## 2026-06-01 14:46:50
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.769.1 (2.892.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.65] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.65] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.65 - packages/orq-rc